Remove Cardano.Config.Parsers
